chore: PR feedbacks

+++ android/src/main/java/com/mobilecoin/mobilecoin_flutter/FfiMobileCoinClient.java
@@ -142,9 +142,7 @@ public class FfiMobileCoinClient {
             AccountKey accountKey = mobileCoinClient.getAccountKey();
             activity.put("balance", balance);
             activity.put("transferableAmount", transferableAmount.getValue().toString());
-            // Use a primitive long so org.json always emits a JSON number.
-            // (avoids any ambiguity around how a wrapped numeric type is serialized).
-            activity.put("blockCount", accountActivity.getBlockCount().longValue());
+            activity.put("blockCount", accountActivity.getBlockCount());
             Set<OwnedTxOut> ownedTxOuts = accountActivity.getAllTokenTxOuts(tokenId).stream()
                 .filter(txOut ->
                     txOut.getReceivedBlockIndex().compareTo(minTxOutBlockIndex) >= 0 ||
